refactor: `entitiy` 패키지명 오타 수정 `entity`로 변경
