+++ app/src/main/java/woowacourse/shopping/data/network/entity/ProductPageEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.network.entitiy
+package woowacourse.shopping.data.network.entity
 
 import woowacourse.shopping.domain.product.ProductSinglePage
