+++ app/src/main/java/woowacourse/shopping/view/main/adapter/recent/RecentProductAdapter.kt
@@ -1,5 +1,6 @@
 package woowacourse.shopping.view.main.adapter.recent
 
+import android.util.Log
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.view.main.adapter.HistoryViewHolder
@@ -9,6 +10,10 @@ class RecentProductAdapter(
     private val items: List<HistoryState>,
     private val handler: HistoryViewHolder.Handler,
 ) : RecyclerView.Adapter<HistoryViewHolder>() {
+    init {
+        Log.d("dsadasas", "init")
+    }
+
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int,
